Fix for Strava API change

+++ shared/api/src/commonMain/kotlin/com.jdamcd.runlog.shared.api/Models.kt
@@ -89,5 +89,5 @@ data class ApiActivityStats(
 data class ApiActivityTotal(
     val count: Int,
     val distance: Float,
-    val moving_time: Int
+    val moving_time: Double
 )
