+++ shared/main/src/commonMain/kotlin/com/jdamcd/runlog/shared/profile/AthleteMapper.kt
@@ -43,7 +43,7 @@ internal class AthleteMapper(private val clock: Clock) {
         lastUpdated = clock.now().epochSeconds
     )
 
-    private fun pace(totals: ApiActivityTotal) = calculatePace(totals.distance, totals.moving_time)
+    private fun pace(totals: ApiActivityTotal) = calculatePace(totals.distance, totals.moving_time.toInt())
 
     fun dbToUi(athlete: AthleteWithStats) = AthleteProfile(
         id = athlete.id,
